LayoutTests/webcl/tests/webcl-test-event-expected.txt

Just testing if a WebCLEvent is working properly...

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS nativeEvent instanceof WebCLEvent is true
PASS nativeEvent instanceof WebCLEvent is true
PASS nativeEvent instanceof WebCLEvent is true
PASS nativeEvent instanceof WebCLEvent is true
PASS count === executionStatus.length is true
PASS successfullyParsed is true

TEST COMPLETE
